refactor inference files and folders
