+++ recipes/AudioMNIST/diffusion/hparams/train.yaml
@@ -207,7 +207,7 @@ diffusion_sample_channels: !ref <diffusion_channels>
 epoch_counter: !new:speechbrain.utils.epoch_loop.EpochCounter
     limit: !ref <number_of_epochs>
 
-vocoder: !name:speechbrain.pretrained.interfaces.HIFIGAN.from_hparams
+vocoder: !name:speechbrain.inference.vocoders.HIFIGAN.from_hparams
     source: !ref <vocoder_model>
