--- speechbrain/pretrained/__init__.py
@@ -1,3 +1,0 @@
-"""Pretrained models"""
-
-from .interfaces import *  # noqa
