+++ speechbrain/utils/parameter_transfer.py
@@ -11,7 +11,7 @@ Authors
 import logging
 import pathlib
 from speechbrain.utils.distributed import run_on_main
-from speechbrain.pretrained.fetching import fetch, FetchFrom, FetchSource
+from speechbrain.utils.fetching import fetch, FetchFrom, FetchSource
 from speechbrain.utils.checkpoints import (
     DEFAULT_LOAD_HOOKS,
     DEFAULT_TRANSFER_HOOKS,
