+++ speechbrain/utils/text_to_sequence.py
@@ -324,12 +324,12 @@ def _g2p_keep_punctuations(g2p_model, text):
     """do grapheme to phoneme and keep the punctuations between the words
     Arguments
     ---------
-    g2p_model: speechbrain pretrained g2p model
+    g2p_model: speechbraininference.text g2p model
     text: string
         the input text
     Example
     -------
-    >>> from speechbrain.pretrained import GraphemeToPhoneme
+    >>> from speechbrain.inference.text import GraphemeToPhoneme
     >>> g2p_model = GraphemeToPhoneme.from_hparams("speechbrain/soundchoice-g2p")
     >>> from speechbrain.utils.text_to_sequence import _g2p_keep_punctuations
     >>> text = "Hi, how are you?"
